+++ MatrixSDK/Contrib/Swift/Data/MXRoom.swift
@@ -1,5 +1,6 @@
 /*
  Copyright 2017 Avery Pierce
+ Copyright 2019 The Matrix.org Foundation C.I.C
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -471,13 +472,14 @@ public extension MXRoom {
      Join this room where the user has been invited.
      
      - parameters:
+        - viaServers The server names to try and join through in addition to those that are automatically chosen
          - completion: A block object called when the operation completes.
          - response: Indicates whether the operation was a success or failure.
      
      - returns: a `MXHTTPOperation` instance.
      */
-    @nonobjc @discardableResult func join(completion: @escaping (_ response: MXResponse<Void>) -> Void) -> MXHTTPOperation {
-        return __join(currySuccess(completion), failure: curryFailure(completion))
+    @nonobjc @discardableResult func join(viaServers: [String]? = nil, completion: @escaping (_ response: MXResponse<Void>) -> Void) -> MXHTTPOperation {
+        return __join(viaServers: viaServers, success:currySuccess(completion), failure: curryFailure(completion))
     }
     
     /**