+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes in Matrix iOS SDK in 0.12.6 (2019-05-)
 Improvements:
  * MXHTTPClient: support multiple SSL pinning modes (none/public key/certificate)
  * MXHTTPClient: Enable the certificate pinning mode by default as soon as some certificates are present in the application bundle.
+ * Join Room: Support via parameters to better handle federation (vector-im/riot-ios/issues/2547).
  * MXEvent: Create a MXEventUnsignedData model for `MXEvent.unsignedData`.
  * MXEvent: Add relatesTo property.
  * Aggregations: Create MXSession.MXAggregations to manage Matrix aggregations API.
@@ -19,6 +20,7 @@ Bug Fix:
 
 API break:
  * MXEvent: unsignedData is now of type MXEventUnsignedData.
+ * MXSession, MXSession, MXRestClient: Add viaServers parameters to all joinRoom methods.
 
 Changes in Matrix iOS SDK in 0.12.5 (2019-05-03)
 ===============================================